requirements.txt
beautifulsoup4
lxml
PyGithub
yattag